// src/app/core/models/memberDocumentType.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export type MemberDocumentType = typeof MemberDocumentType[keyof typeof MemberDocumentType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MemberDocumentType = {
  IdentificationDocument: 1,
  ProofOfAddress: 2,
  MarriageCertificate: 3,
  Passport: 4,
  WorkPermit: 5,
  BirthCertificate: 6,
  DeathCertificate: 7,
  BankingDocument: 8,
  Other: 99,
} as const;